pyproject.toml
[tool.poetry]
name = "bar-maker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bar_maker"}]

[tool.poetry.dependencies]
python = "^3.9"
gcsfs = "^2022.8.2"
s3fs = "^2022.8.2"
adlfs = "^2022.7.0"
polars = "^0.14.8"
black = "^22.8.0"
isort = "^5.10.1"
python-dotenv = "^0.21.0"
numpy = "^1.23.2"
loguru = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
